+++ mayan/apps/documents/apps.py
@@ -161,7 +161,7 @@ class DocumentsApp(MayanAppConfig):
             label=_('Create a document type'),
             description=_(
                 'Every uploaded document must be assigned a document type, '
-                'it is the basic way Mayan EDMS categorizes documents.'
+                'it is the basic way PROGED EDMS categorizes documents.'
             ), condition=lambda: not DocumentType.objects.exists(),
             view='documents:document_type_list'
         )